Fix bug for read_vcf function in geno_read_base.R
There's an error when loading vcf data into Bioflow. 
Genotype calls that should be 2 are being called as 1. 
Genotype calls that should be 1 are being called as 0. 
The functions were splitting genotype strings without removing the separator (e.g., "A/G"), so the second “allele” becomes "/G" and never equals "G"